Move userAgent func to options.go
